+++ src/App.tsx
@@ -2,14 +2,14 @@ import React from 'react'
 import { Helmet } from 'react-helmet'
 import 'twin.macro'
 import tw from 'twin.macro'
-import GlobalStyles from './GlobalStyles'
-import WideView from './components/WideView'
+import GlobalStyles from '$/GlobalStyles'
+import WideView from '$/components/WideView'
 import NarrowView from '$/components/NarrowView'
 import { Container } from 'toasterhea'
-import { Layer } from './types'
-import { useUiStore } from './stores/ui'
-import { useAutoConnectEffect } from './stores/data'
-import Debug from './components/Debug'
+import { Layer } from '$/shared/types'
+import { useUiStore } from '$/stores/ui'
+import { useAutoConnectEffect } from '$/stores/data'
+import Debug from '$/components/Debug'
 
 const App = () => {
     const { theme } = useUiStore()
