+++ src/components/WideView.tsx
@@ -1,9 +1,9 @@
 import Clock from '$/components/ui/Clock'
 import Controller from '$/components/ui/Controller'
 import WaterBar from '$/components/ui/WaterBar'
-import { Layer } from '$/types'
+import { Layer } from '$/shared/types'
 import { useDataStore, useMajorState, useWaterLevel } from '$/stores/data'
-import { Prop } from '$/types'
+import { Prop } from '$/shared/types'
 import { mlToL } from '$/utils'
 import React, { HTMLAttributes, ReactNode } from 'react'
 import { toaster } from 'toasterhea'
